requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
-e .
    # via -r requirements.in
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer-slim
cloudpickle==3.0.0
    # via scalene
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.53.1
    # via matplotlib
h5py==3.11.0
    # via -r requirements.in
idna==3.7
    # via requests
imageio==2.34.2
    # via scikit-image
jinja2==3.1.4
    # via scalene
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
llvmlite==0.43.0
    # via numba
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
networkx==3.3
    # via scikit-image
numba==0.60.0
    # via -r requirements.in
numpy==2.0.0
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   imageio
    #   matplotlib
    #   numba
    #   pandas
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
nvidia-ml-py==12.555.43
    # via scalene
nvidia-ml-py3==7.352.0
    # via scalene
packaging==24.1
    # via
    #   lazy-loader
    #   matplotlib
    #   pooch
    #   scikit-image
pandas==2.2.2
    # via -r requirements.in
pillow==10.4.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
platformdirs==4.2.2
    # via pooch
pooch==1.8.2
    # via scikit-image
psutil==6.0.0
    # via scalene
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
requests==2.32.3
    # via pooch
rich==13.7.1
    # via scalene
scalene==1.5.42.2
    # via -r requirements.in
scikit-image[data]==0.24.0
    # via -r requirements.in
scikit-learn==1.5.1
    # via -r requirements.in
scipy==1.14.0
    # via
    #   scikit-image
    #   scikit-learn
six==1.16.0
    # via python-dateutil
threadpoolctl==3.5.0
    # via scikit-learn
tifffile==2024.7.2
    # via scikit-image
typer-slim==0.12.3
    # via -r requirements.in
typing-extensions==4.12.2
    # via typer-slim
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
wheel==0.43.0
    # via scalene
